Move moderation commands cog to data archive
